Make Server Send Frames
Make the server send frames of a movie file. May also need to implement threading so that there is an RTP packet handler and RTSP packet handler. Also might be worth it to fully implement the RTP and RTSP packet structure, but this is a simplified version that should accomplish the same thing.